tools/patches/conan/libjpeg-turbo/conan_requirements.txt
# Generated for conan_package_tools==0.31.1 bincrafters_package_tools==0.22.0 conan==1.22.2 
appdirs==1.4.3
bincrafters-package-tools==0.22.0
bottle==0.12.18
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
conan==1.22.2
conan-package-tools==0.31.1
deprecation==2.0.7
distlib==0.3.0
distro==1.1.0
fasteners==0.15
filelock==3.0.12
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
node-semver==0.6.1
packaging==20.1
patch-ng==1.17.2
pluginbase==0.7
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
six==1.12.0
tabulate==0.8.2
tqdm==4.43.0
urllib3==1.25.8
virtualenv==20.0.4
zipp==2.2.0